fixes matter idl test
