+++ scripts/py_matter_idl/matter_idl/tests/outputs/large_lighting_app/cpp-app/cluster-init-callback.cpp
@@ -1,0 +1,95 @@
+#include <app-common/zap-generated/callback.h>
+#include <app-common/zap-generated/ids/Clusters.h>
+#include <lib/support/Span.h>
+#include <protocols/interaction_model/Constants.h>
+
+using namespace chip;
+
+// Cluster Init Functions
+void emberAfClusterInitCallback(EndpointId endpoint, ClusterId clusterId)
+{
+    switch (clusterId)
+    {
+    case app::Clusters::AccessControl::Id:
+        emberAfAccessControlClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::AdministratorCommissioning::Id:
+        emberAfAdministratorCommissioningClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::BasicInformation::Id:
+        emberAfBasicInformationClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::ColorControl::Id:
+        emberAfColorControlClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::Descriptor::Id:
+        emberAfDescriptorClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::DiagnosticLogs::Id:
+        emberAfDiagnosticLogsClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::EthernetNetworkDiagnostics::Id:
+        emberAfEthernetNetworkDiagnosticsClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::FixedLabel::Id:
+        emberAfFixedLabelClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::GeneralCommissioning::Id:
+        emberAfGeneralCommissioningClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::GeneralDiagnostics::Id:
+        emberAfGeneralDiagnosticsClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::GroupKeyManagement::Id:
+        emberAfGroupKeyManagementClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::Groups::Id:
+        emberAfGroupsClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::Identify::Id:
+        emberAfIdentifyClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::LevelControl::Id:
+        emberAfLevelControlClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::LocalizationConfiguration::Id:
+        emberAfLocalizationConfigurationClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::NetworkCommissioning::Id:
+        emberAfNetworkCommissioningClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::OccupancySensing::Id:
+        emberAfOccupancySensingClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::OnOff::Id:
+        emberAfOnOffClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::OperationalCredentials::Id:
+        emberAfOperationalCredentialsClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::OtaSoftwareUpdateRequestor::Id:
+        emberAfOtaSoftwareUpdateRequestorClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::SoftwareDiagnostics::Id:
+        emberAfSoftwareDiagnosticsClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::Switch::Id:
+        emberAfSwitchClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::ThreadNetworkDiagnostics::Id:
+        emberAfThreadNetworkDiagnosticsClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::TimeFormatLocalization::Id:
+        emberAfTimeFormatLocalizationClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::UserLabel::Id:
+        emberAfUserLabelClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::WiFiNetworkDiagnostics::Id:
+        emberAfWiFiNetworkDiagnosticsClusterInitCallback(endpoint);
+        break;
+    default:
+        // Unrecognized cluster ID
+        break;
+    }
+}
