+++ scripts/py_matter_idl/matter_idl/tests/outputs/several_clusters/cpp-app/callback-stub.cpp
@@ -1,23 +1,6 @@
 #include <app-common/zap-generated/callback.h>
-#include <app-common/zap-generated/ids/Clusters.h>
-#include <lib/support/Span.h>
-#include <protocols/interaction_model/Constants.h>
 
 using namespace chip;
-
-// Cluster Init Functions
-void emberAfClusterInitCallback(EndpointId endpoint, ClusterId clusterId)
-{
-    switch (clusterId)
-    {
-    case app::Clusters::Third::Id:
-        emberAfThirdClusterInitCallback(endpoint);
-        break;
-    default:
-        // Unrecognized cluster ID
-        break;
-    }
-}
 void __attribute__((weak)) emberAfThirdClusterInitCallback(EndpointId endpoint)
 {
     // To prevent warning
